+++ metric/mean_squared_error_test.go
@@ -11,12 +11,12 @@ func Test_MeanSquaredError_Value(t *testing.T) {
 	var (
 		predictions *matrix.Matrix
 		targets     *matrix.Matrix
-		got         float64
+		got         float32
 		err         error
 	)
 
-	predictions = mustMatrix(t, 2, 2, []float64{1, 2, 3, 4})
-	targets = mustMatrix(t, 2, 2, []float64{1.5, 1, 2, 5})
+	predictions = mustMatrix(t, 2, 2, []float32{1, 2, 3, 4})
+	targets = mustMatrix(t, 2, 2, []float32{1.5, 1, 2, 5})
 
 	got, err = metric.MeanSquaredError{}.Value(predictions, targets)
 	if err != nil {
@@ -30,12 +30,12 @@ func Test_MeanSquaredError_ValidatesShape(t *testing.T) {
 	var (
 		predictions *matrix.Matrix
 		targets     *matrix.Matrix
-		got         float64
+		got         float32
 		err         error
 	)
 
-	predictions = mustMatrix(t, 1, 2, []float64{1, 2})
-	targets = mustMatrix(t, 2, 1, []float64{1, 2})
+	predictions = mustMatrix(t, 1, 2, []float32{1, 2})
+	targets = mustMatrix(t, 2, 1, []float32{1, 2})
 
 	got, err = metric.MeanSquaredError{}.Value(predictions, targets)
 	if err == nil {
