+++ model/train_metrics.go
@@ -3,5 +3,5 @@ package model
 // TrainMetrics reports metrics from one training batch.
 type TrainMetrics struct {
 	// Loss is the scalar loss computed before the optimizer update.
-	Loss float64
+	Loss float32
 }
